yfinance_db/requirements.txt
yfinance
pandas
SQLAlchemy
psycopg2-binary
python-dotenv